% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightRegression-package.R
\docType{package}
\name{iNZightRegression-package}
\alias{iNZightRegression}
\alias{iNZightRegression-package}
\title{iNZightRegression: Tools for Exploring Regression Models with 'iNZight'}
\description{
Provides a suite of functions to use with regression models, including summaries, residual plots, and factor comparisons. Used as part of the Model Fitting module of 'iNZight', a graphical user interface providing easy exploration and visualisation of data for students of statistics, available in both desktop and online versions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://inzightvit.github.io/iNZightRegression/}
  \item Report bugs at \url{https://github.com/iNZightVIT/iNZightRegression/issues}
}

}
\author{
\strong{Maintainer}: Tom Elliott \email{tom.elliott@auckland.ac.nz} (\href{https://orcid.org/0000-0002-7815-6318}{ORCID})

Authors:
\itemize{
  \item Ken Deng
  \item Simon Potter
  \item David Banks
}

Other contributors:
\itemize{
  \item Danny Chang [contributor]
}

}
\keyword{internal}
