% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightRegression.R
\docType{package}
\name{iNZightRegression}
\alias{iNZightRegression}
\title{Additional functionality for iNZight}
\description{
This package provides some additional functions to be used when
fitting regression models. It supports linear, general linear, and
survey (general) linear models.
}
\author{
Simon Potter, David Banks, Tom Elliott.
}
