% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzsummary.R, R/inzsummary_lm.R
\name{inzsummary}
\alias{inzsummary}
\alias{inzsummary.lm}
\title{inzsummary method}
\usage{
\method{inzsummary}{lm}(x, ..., env = parent.frame())
}
\arguments{
\item{x}{an \code{lm}, \code{glm}, or \code{svyglm} object}

\item{...}{additional arguments passed to \code{iNZightSummary}}

\item{env}{the environment for evaluating things (e.g., bootstraps)}
}
\value{
An object of class \code{summary.lm}, \code{summary.glm}, or
\code{summary.svyglm}.
}
\description{
inzsummary method

Summary method for linear models
}
\seealso{
\link{iNZightSummary}
}
