% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses.moecalc.R
\name{ses.moecalc}
\alias{ses.moecalc}
\title{Constructor function for Standard Error Class}
\usage{
ses.moecalc(ses, ses.diffs)
}
\arguments{
\item{ses}{numeric, vector of standard errors.}

\item{ses.diffs}{matrix of standard error differences.}
}
\value{
An object of class \code{ses.moecalc}.
}
\description{
Constructs a standard error object, consisting of standard errors and
a difference matrix. Used in margin of error calculations, see
\code{\link{moecalc}}.
}
\seealso{
\code{\link{moecalc}}, \code{\link{multicomp}}
}
\author{
David Banks.
}
