\name{allPartialResPlots}
\alias{allPartialResPlots}
\title{
  Draws multiple partial residual plots
}
\description{
  Given a fitted model, this function draws partial residual plots for
  all continuous explanatory variables in the model. Clicking the device
  displays the next plot in the sequence.
}
\usage{
allPartialResPlots(fit)
}

\arguments{
  \item{fit}{
    An lm or glm object
  }
}
\value{
  None.
}
\author{
David Banks
}