% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comp.R
\name{compare_models}
\alias{compare_models}
\title{Compare regression models using AIC, BIC, etc}
\usage{
compare_models(x, ...)
}
\arguments{
\item{x}{a regression model (lm, glm, svyglm, ...)}

\item{...}{other models}
}
\value{
an matrix object
}
\description{
Obtain a quick model comparison matrix for a selection of models
}
\author{
Tom Elliott
}
