% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorMeans.R
\name{factorComp}
\alias{factorComp}
\title{Compare factor levels}
\usage{
factorComp(fit, factor)
}
\arguments{
\item{fit}{a lm/glm/svyglm object}

\item{factor}{the name of the factor to compare}
}
\value{
a factor level comparison matrix with estimates, CIs, 
        and (adjusted) p-values
}
\description{
Computes confidence intervals for the pairwise differences between levels 
of a factor, based off of \link{TukeyHSD}.
}
\author{
Tom Elliott
}
