% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzplot.R
\name{inzplot.lm}
\alias{inzplot.lm}
\title{Diagnostic Plots for Regression Models}
\usage{
\method{inzplot}{lm}(
  x,
  which = c("residual", "scale-location", "leverage", "cooks", "normal", "hist"),
  show.bootstraps = TRUE,
  ...,
  env = parent.frame()
)
}
\arguments{
\item{which}{the type of plot to draw}

\item{...}{additional arguments}

\item{env}{the environment for evaluating things (e.g., bootstraps)}

\item{fit}{a regression model}
}
\value{
A ggplot object
}
\description{
Diagnostic Plots for Regression Models
}
\section{Plot types}{

There are several plot types available:
\itemize{
\item residual versus fitted
\item scale-location
\item residual versus leverage
\item Cook's distance
\item normal Q-Q
\item histogram
}
}

\examples{
iris_fit <- lm(Sepal.Width ~ Sepal.Length, data = iris)
inzplot(iris_fit)
}
\author{
Tom Elliott
}
