% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inzsummary.R
\name{inzsummary}
\alias{inzsummary}
\alias{inzsummary.lm}
\title{inzsummary method}
\usage{
\method{inzsummary}{lm}(
  x,
  method = c("standard", "bootstrap"),
  reorder.factors = FALSE,
  digits = max(3, getOption("digits") - 3),
  symbolic.cor = x$symbolic.cor,
  signif.stars = getOption("show.signif.stars"),
  exclude = NULL,
  exponentiate.ci = FALSE,
  ...,
  env = parent.frame()
)
}
\arguments{
\item{x}{an object of class \code{"lm"}, \code{"glm"} or \code{"svyglm"}, usually the result of a call to the corresponding function.}

\item{method}{the type of inference to perform: "standard" uses Normal theory, "bootstrap" uses bootstrap resampling.}

\item{reorder.factors}{if \code{FALSE} (default), factor order is left unchanged, if \code{TRUE}, the most common level is set as the baseline. Alternatively can be a character value of a level to use as the baseline.}

\item{digits}{the number of significant digits to display in results}

\item{symbolic.cor}{logical, if \code{TRUE}, print the correlations in a symbolic form (see \code{\link{symnum}}), rather than as numbers.}

\item{signif.stars}{logical, if \code{TRUE}, \sQuote{significance stars} are printed for each coefficient.}

\item{exclude}{a character vector of names of variables to be excluded from the summary output (i.e., confounding variables).}

\item{exponentiate.ci}{logical, if \code{TRUE}, the exponential of the confidence intervals will be printed if appropriate (log/logit link or log transformed response)}

\item{...}{additional arguments, ignored.}

\item{env}{the environment for evaluating things (e.g., bootstraps)}
}
\value{
Summary information is printed to the console.
}
\description{
Displays a summary of a regression model fitted using \code{lm}, \code{glm}, or \code{survey::svyglm}.
}
\author{
Tom Elliott
}
