% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/moecalc.R
\name{moecalc}
\alias{moecalc}
\title{Margin of Error Calculations}
\usage{
moecalc(x, factorname = NULL, levelnames = NULL, coef.idx = NULL,
  est = NULL, ci = NULL, base = TRUE, basename = "base",
  conf.level = 1.96)
}
\arguments{
\item{x}{a model object (i.e. \code{lm}, \code{glm}, \code{svyglm}), or an
\code{ses.moecalc} object.}

\item{factorname}{character, the name of the factor. Can be an interaction.}

\item{levelnames}{the names of each level of the factor.}

\item{coef.idx}{the coefficient index.}

\item{est}{estimate.}

\item{ci}{the relevant confidence interval; only use when the length of
\code{est} is 2. Default is 1.96, default 95\% level.}

\item{base}{logical, whether a baseline level is set. defaults to \code{TRUE}.}

\item{basename}{label name to apply to baseline level if \code{base = TRUE}.}

\item{conf.level}{confidence level for both error bars and confidence intervals.}
}
\value{
An \code{moecalc} object, which has the following methods:
\code{plot}, \code{print}, and \code{summary}.

\code{ses.moecalc} produces an object of class \code{"ses.moecalc"}.
}
\description{
...
}
\author{
Danny Chang, David Banks
}

