% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moe-multicomp.R
\name{multicomp}
\alias{multicomp}
\title{Multiple Comparisons Between Factor Levels}
\usage{
multicomp(x, ...)
}
\arguments{
\item{x}{an \code{moecalc} object.}

\item{...}{additional parameters to be passed to the function (currently
ignored).}
}
\description{
This is a method for \code{moecalc} objects. Most other \code{moecalc}
methods are concerned with how each factor level differs from the
baseline. This does not show the significance of differences between
non-baseline levels of a factor.
\cr\cr
Currently the p-values are not adjusted. As a result, more comparisons
are likely to be significantly different when they shouldn't be.
}
\author{
Simon Potter.
}

