% Generated by roxygen2 (4.0.2): do not edit by hand
\name{partialResPlot}
\alias{partialResPlot}
\title{Draw multiple partial residual plots}
\usage{
partialResPlot(fit, varname, showBootstraps = nrow(fit$model) >= 30 &
  nrow(fit$model) < 4000)
}
\arguments{
\item{fit}{an \code{lm}, \code{glm} or \code{svyglm} object.}

\item{varname}{}

\item{showBootstraps}{}
}
\value{
None.
}
\description{
This function draws partial residual plots for all continuous
explanatory variables in a given model. Clicking the device
displays the next plot in the sequence.
}
\author{
David Banks, Tom Elliott.
}
\seealso{
\code{\link{partialResPlot}}
}

