% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialResidual.R
\name{partialResPlot}
\alias{partialResPlot}
\title{Draw multiple partial residual plots}
\usage{
partialResPlot(fit, varname, showBootstraps = nrow(fit$model) >= 30 &
  nrow(fit$model) < 4000, use.inzightplots = FALSE,
  env = parent.frame())
}
\arguments{
\item{fit}{an \code{lm}, \code{glm} or \code{svyglm} object.}

\item{varname}{explanatory variable}

\item{showBootstraps}{logical. If there are more than than or
equal to 30 and less than 4000 samples in the model, plot
bootstrap smoothers, by default.}

\item{use.inzightplots}{logical, if \code{TRUE}, the iNZightPlots package will be used for
plotting.}

\item{env}{environment where the data is stored for bootstrapping}
}
\value{
None.
}
\description{
This function draws partial residual plots for all continuous
explanatory variables in a given model. Clicking the device
displays the next plot in the sequence.
}
\seealso{
\code{\link{partialResPlot}}
}
\author{
David Banks, Tom Elliott.
}
